Initial setup with UE4 and GitHub Desktop
feat: Added in the base of the application

This is a starting example of combining the UE4 VR template with level
transitions. The idea is to combine this with LightYourWill so that we
can have a VR lobby and users can choose whichever training levels they
would like to practice or play. Support is still with only DK2 and the
infrared camera or CV1 with the Oculus camera. No support for Vive and
room-scale tracking yet.